### 2021-04-28 3a76c29
add ambient audio collection to checkbox toggles in study creation validation
