+++ config/constants.py
@@ -25,7 +25,7 @@ CHECKBOX_TOGGLES = [
     "gyro",
     "magnetometer",
     "devicemotion",
-    "background_audio",
+    "ambient_audio",
     "reachability",
     "allow_upload_over_cellular_data",
     "use_anonymized_hashing",
