### 2021-05-12 506bd8b
Bump django from 2.2.19 to 2.2.20
Bumps [django](https://github.com/django/django) from 2.2.19 to 2.2.20.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/2.2.19...2.2.20)

Signed-off-by: dependabot[bot] <support@github.com>
