### 2021-05-20 214e273
Prevent error if Forest has no output
