### 2021-05-20 634e1ea
Add ForestTask.forest_output_exists
