+++ database/migrations/0054_foresttask_forest_output_exists.py
@@ -1,0 +1,18 @@
+# Generated by Django 2.2.19 on 2021-05-20 19:19
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('database', '0053_devicesettings_ambient_audio'),
+    ]
+
+    operations = [
+        migrations.AddField(
+            model_name='foresttask',
+            name='forest_output_exists',
+            field=models.NullBooleanField(),
+        ),
+    ]
