### 2021-05-20 683fcfa
Fix ForestTask.process_download_end_time data type
