+++ database/migrations/0055_auto_20210520_1931.py
@@ -1,0 +1,18 @@
+# Generated by Django 2.2.19 on 2021-05-20 19:31
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('database', '0054_foresttask_forest_output_exists'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='foresttask',
+            name='process_download_end_time',
+            field=models.DateTimeField(blank=True, null=True),
+        ),
+    ]
