### 2021-05-24 daad883
Fix bug in summary statistics api
