+++ api/tableau_api/views.py
@@ -74,7 +74,7 @@ class SummaryStatisticDailyStudyView(TableauApiView):
         if order_direction == "descending":
             order_by = "-" + order_by
         # TODO: change to be based participant studies
-        queryset = SummaryStatisticDaily.objects.filter(study__object_id=study_id)
+        queryset = SummaryStatisticDaily.objects.filter(participant__study__object_id=study_id)
         if participant_ids:
             queryset = queryset.filter(participant__patient_id__in=participant_ids)
         if end_date:
