### 2021-05-24 73c401c
Set jasmine/willow task on SummaryStatisticDaily
