### 2021-05-24 2768954
Fix Summary Statistics serializer
