### 2021-05-25 fcdc7f0
Filter chunks based on task data start/end
