+++ services/celery_forest.py
@@ -32,57 +32,57 @@ TREE_TO_FOREST_FUNCTION = {
 
 
 def create_forest_celery_tasks():
-    pending_trackers = ForestTask.objects.filter(status=ForestTask.Status.queued)
+    pending_tasks = ForestTask.objects.filter(status=ForestTask.Status.queued)
 
     # with make_error_sentry(sentry_type=SentryTypes.data_processing):  # add a new type?
     with NullErrorHandler():  # for debugging, does not suppress errors
-        for tracker in pending_trackers:
-            print(f"Queueing up celery task for {tracker.participant} on tree {tracker.forest_tree} from {tracker.data_date_start} to {tracker.data_date_end}")
-            enqueue_forest_task(args=[tracker.id])
+        for task in pending_tasks:
+            print(f"Queueing up celery task for {task.participant} on tree {task.forest_tree} from {task.data_date_start} to {task.data_date_end}")
+            enqueue_forest_task(args=[task.id])
 
 
 #run via celery as long as tasks exist
 @forest_celery_app.task(queue=FOREST_QUEUE)
 def celery_run_forest(forest_task_id):
     with transaction.atomic():
-        tracker = ForestTask.objects.filter(id=forest_task_id).first()
+        task = ForestTask.objects.filter(id=forest_task_id).first()
 
-        participant = tracker.participant
-        forest_tree = tracker.forest_tree
+        participant = task.participant
+        forest_tree = task.forest_tree
         
         # Check if there already is a running task for this participant and tree, handling
         # concurrency and requeuing of the ask if necessary
-        trackers = (
+        tasks = (
             ForestTask
                 .objects
                 .select_for_update()
                 .filter(participant=participant, forest_tree=forest_tree)
         )
-        if trackers.filter(status=ForestTask.Status.running).exists():
-            enqueue_forest_task(args=[tracker.id])
+        if tasks.filter(status=ForestTask.Status.running).exists():
+            enqueue_forest_task(args=[task.id])
             return
         
-        # Get the chronologically earliest tracker that's queued
-        tracker = (
-            trackers
+        # Get the chronologically earliest task that's queued
+        task = (
+            tasks
                 .filter(status=ForestTask.Status.queued)
                 .order_by("-data_date_start")
                 .first()
         )
-        if tracker is None:
+        if task is None:
             return
         
-        # Set metadata on the tracker
-        tracker.status = ForestTask.Status.running
-        tracker.forest_version = get_distribution("forest").version
-        tracker.process_start_time = timezone.now()
-        tracker.save(update_fields=["status", "forest_version", "process_start_time"])
+        # Set metadata on the task
+        task.status = ForestTask.Status.running
+        task.forest_version = get_distribution("forest").version
+        task.process_start_time = timezone.now()
+        task.save(update_fields=["status", "forest_version", "process_start_time"])
 
     try:
         # Save file size data
         # The largest UTC offsets are -12 and +14
-        min_datetime = datetime.combine(tracker.data_date_start, datetime.min.time()) - timedelta(hours=12)
-        max_datetime = datetime.combine(tracker.data_date_end, datetime.max.time()) + timedelta(hours=14)
+        min_datetime = datetime.combine(task.data_date_start, datetime.min.time()) - timedelta(hours=12)
+        max_datetime = datetime.combine(task.data_date_end, datetime.max.time()) + timedelta(hours=14)
         chunks = (
             ChunkRegistry
                 .objects
@@ -90,40 +90,40 @@ def celery_run_forest(forest_task_id):
[…]
