### 2021-05-25 380ef2d
Handle no data case
