### 2021-05-25 2a01b49
Add unique constraint to SummaryStatisticDaily and fix association of task
