+++ database/migrations/0057_auto_20210525_1920.py
@@ -1,0 +1,28 @@
+# Generated by Django 2.2.19 on 2021-05-25 19:20
+
+from django.db import migrations, models
+import django.db.models.deletion
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('database', '0056_auto_20210524_2128'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='jasmine_task',
+            field=models.ForeignKey(blank=True, null=True, on_delete=django.db.models.deletion.PROTECT, related_name='jasmine_summary_statistics', to='database.ForestTask'),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='willow_task',
+            field=models.ForeignKey(blank=True, null=True, on_delete=django.db.models.deletion.PROTECT, related_name='willow_summary_statistics', to='database.ForestTask'),
+        ),
+        migrations.AddConstraint(
+            model_name='summarystatisticdaily',
+            constraint=models.UniqueConstraint(fields=('date', 'participant'), name='unique_summary_statistic'),
+        ),
+    ]
