### 2021-05-25 18d0640
Fix bug in summary statistics parsing
