### 2021-05-25 bb43945
Convert jasmine IntegerField to FloatField
