+++ database/migrations/0058_auto_20210525_1929.py
@@ -1,0 +1,53 @@
+# Generated by Django 2.2.19 on 2021-05-25 19:29
+
+from django.db import migrations, models
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('database', '0057_auto_20210525_1920'),
+    ]
+
+    operations = [
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='distance_diameter',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='distance_from_home',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='distance_traveled',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='flight_distance_average',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='flight_distance_standard_deviation',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='flight_duration_average',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='flight_duration_standard_deviation',
+            field=models.FloatField(blank=True, null=True),
+        ),
+        migrations.AlterField(
+            model_name='summarystatisticdaily',
+            name='home_duration',
+            field=models.FloatField(blank=True, null=True),
+        ),
+    ]
