+++ database/tableau_api_models.py
@@ -279,15 +279,15 @@ class ForestTask(TimestampedModel):
 class SummaryStatisticDaily(TimestampedModel):
     participant = models.ForeignKey(Participant, on_delete=models.CASCADE)
     date = models.DateField(db_index=True)
-    distance_diameter = models.IntegerField(null=True, blank=True)
-    distance_from_home = models.IntegerField(null=True, blank=True)
-    distance_traveled = models.IntegerField(null=True, blank=True)
-    flight_distance_average = models.IntegerField(null=True, blank=True)
-    flight_distance_standard_deviation = models.IntegerField(null=True, blank=True)
-    flight_duration_average = models.IntegerField(null=True, blank=True)
-    flight_duration_standard_deviation = models.IntegerField(null=True, blank=True)
+    distance_diameter = models.FloatField(null=True, blank=True)
+    distance_from_home = models.FloatField(null=True, blank=True)
+    distance_traveled = models.FloatField(null=True, blank=True)
+    flight_distance_average = models.FloatField(null=True, blank=True)
+    flight_distance_standard_deviation = models.FloatField(null=True, blank=True)
+    flight_duration_average = models.FloatField(null=True, blank=True)
+    flight_duration_standard_deviation = models.FloatField(null=True, blank=True)
     gps_data_missing_duration = models.IntegerField(null=True, blank=True)
-    home_duration = models.IntegerField(null=True, blank=True)
+    home_duration = models.FloatField(null=True, blank=True)
     physical_circadian_rhythm = models.FloatField(null=True, blank=True)
     physical_circadian_rhythm_stratified = models.FloatField(null=True, blank=True)
     radius_of_gyration = models.IntegerField(null=True, blank=True)
