### 2021-06-04 6eee8b2
Bump django from 2.2.19 to 2.2.21
Bumps [django](https://github.com/django/django) from 2.2.19 to 2.2.21.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/2.2.19...2.2.21)

---
updated-dependencies:
- dependency-name: django
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
