### 2021-06-10 12189a7
Hide inactive SummaryStatisticDaily fields
