+++ api/tableau_api/constants.py
@@ -1,20 +1,12 @@
 from django.db.models.fields import IntegerField, FloatField, DateTimeField, DateField, BooleanField
 
 SERIALIZABLE_FIELD_NAMES = [
-    "acceleration_direction",
-    "accelerometer_coverage_fraction",
-    "accelerometer_signal_variability",
-    "accelerometer_univariate_summaries",
-    "awake_onset_time",
-    "call_incoming_count",
-    "call_incoming_degree",
-    "call_incoming_duration",
-    "call_incoming_responsiveness",
-    "call_outgoing_count",
-    "call_outgoing_degree",
-    "call_outgoing_duration",
+    # Metadata
     "date",
-    "device_proximity",
+    "participant_id",
+    "study_id",
+    
+    # GPS
     "distance_diameter",
     "distance_from_home",
     "distance_traveled",
@@ -24,27 +16,48 @@ SERIALIZABLE_FIELD_NAMES = [
     "flight_duration_standard_deviation",
     "gps_data_missing_duration",
     "home_duration",
-    "participant_id",
-    "physical_circadian_rhythm",
-    "physical_circadian_rhythm_stratified",
-    "radius_of_gyration",
-    "significant_location_count",
-    "significant_location_entropy",
-    "sleep_duration",
-    "sleep_onset_time",
-    "stationary_fraction",
-    "study_id",
-    "text_incoming_count",
-    "text_incoming_degree",
-    "text_incoming_length",
-    "text_incoming_responsiveness",
-    "text_outgoing_count",
-    "text_outgoing_degree",
-    "text_outgoing_length",
-    "text_reciprocity",
-    "total_power_events",
-    "total_screen_events",
-    "total_unlock_events",
+    # "physical_circadian_rhythm",
+    # "physical_circadian_rhythm_stratified",
+    # "radius_of_gyration",
+    # "significant_location_count",
+    # "significant_location_entropy",
+    # "stationary_fraction",
+    
+    # Texts
+    # "text_incoming_count",
+    # "text_incoming_degree",
+    # "text_incoming_length",
+    # "text_incoming_responsiveness",
+    # "text_outgoing_count",
+    # "text_outgoing_degree",
+    # "text_outgoing_length",
+    # "text_reciprocity",
+    
+    # Calls
+    # "call_incoming_count",
+    # "call_incoming_degree",
+    # "call_incoming_duration",
+    # "call_incoming_responsiveness",
+    # "call_outgoing_count",
+    # "call_outgoing_degree",
+    # "call_outgoing_duration",
+    
+    # Accelerometer
+    # "acceleration_direction",
+    # "accelerometer_coverage_fraction",
+    # "accelerometer_signal_variability",
+    # "accelerometer_univariate_summaries",
+    # "device_proximity",
+    
+    # Power state
+    # "total_power_events",
+    # "total_screen_events",
+    # "total_unlock_events",
+    
+    # Multiple domains
+    # "awake_onset_time",
+    # "sleep_duration",
+    # "sleep_onset_time",
 ]
 
 SERIALIZABLE_FIELD_NAMES_DROPDOWN = [(f, f) for f in SERIALIZABLE_FIELD_NAMES]
