+++ database/migrations/0053_foresttask_forest_output_exists.py
@@ -6,7 +6,7 @@ from django.db import migrations, models
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('database', '0053_devicesettings_ambient_audio'),
+        ('database', '0052_foresttask_params_dict_cache'),
     ]
 
     operations = [
