+++ requirements_data_processing.txt
@@ -2,4 +2,4 @@
 celery==4.4.7
 
 # This is temporary until forest is open-sourced
-git+ssh://git@github.com/onnela-lab/forest@zagaran-output-pattern-changes
+git+https://git@github.com/onnela-lab/forest
