### 2021-06-25 #3216 @fboehm
@erik-whiting - would you be willing to review this submission to JOSS?
