### 2021-06-25 #3417 @whedon
[REVIEW]: Beiwe: A data collection platform for high-throughput digital phenotyping
**Submitting author:** @jponnela (<a href="http://orcid.org/0000-0001-6613-8668">Jukka-Pekka Onnela</a>)
**Repository:** <a href="https://github.com/onnela-lab/beiwe-backend" target ="_blank">https://github.com/onnela-lab/beiwe-backend</a>
**Version:** v1.0.0
