### 2021-06-26 #3417 @erik-whiting
hi @fboehm , quick question for you. Do I just start making issues/comments on their repo as I go through review or should I make all issues at one time? Also, do I need to introduce myself to them or anything? I've never done a none-blind review before so just want to make sure I'm up-to-date on procedure and etiquette.
