### 2021-06-27 #3417 @fboehm
hi, @erik-whiting - thank you for the questions. I suggest that you make comments and open issues as you proceed through the review checklist. There's no need to wait for everything at once. I also suggest that you link from this "review" thread to any issues that you open on the submission repository, ie, the beiwe repo. As long as you do that, I think that no introduction is needed. Thanks again for asking these questions. I know that the JOSS review format can feel a little foreign at first.
