### 2021-06-28 onnela-lab/beiwe-backend#208 @erik-whiting
Peer Review - Paper Issues
Some issues from my review of the paper. Review is being coordinated [here](https://github.com/openjournals/joss-reviews/issues/3417)

_**Summary:**_ - This is good and ticks all the boxes but I have one note. Are the specs on line 11 - 16 necessary or in the right place? It seems like it interrupts the sentences before and after.

_**Statement of Need**_ - This ticks most of the boxes however I feel like the statement "Many researchers are increasingly advocating a more substantial role for large scale phenotyping as a route to advances in the biomedical sciences" is a claim needing a supporting source.

_**State of the Field**_ - The checklist question for this section is "Do the authors describe how this software compares to other commonly-used packages?" This part seems to be missing or could use some elaboration. You say in the Background subsection that surveys are typically used to study human behavior; is this the closest thing to an alternative to Beiwe in your field? Is there no software kind of like Beiwe? If not, can you make that more explicit?

On a related note, you say Beiwe's collection of raw sensor data is more accurate than the data that could be gathered from an SDK (I assume you mean things like React Native, Xamarin, etc., right?). This leads me to believe there are systems out there collecting data via SDKs . Do these systems exist? If so, can you mention them in the spirit of comparing your software to "other commonly-used packages"?

_**Quality of Writing**_ - The writing quality is excellent, nice work.

_**References**_ - References are good except for my note about citing support for your claim in the Background section (lines 30 and 31)

Tag me if you have any questions, updates, or disagreements! I will update the issue and review checklist accordingly.
