### 2021-06-28 #3417 @erik-whiting
I opened [this issue](https://github.com/onnela-lab/beiwe-backend/issues/208) as my review to their paper
