### 2021-06-29 onnela-lab/beiwe-backend#208 @Tam-Pham
I very much agree with @erik-whiting's point regarding the ***State of the Field***. To my knowledge, [Purple Robot](https://github.com/cbitstech/Purple-Robot) is another platform for real-time mobile phone sensor data acquisition. I believe it would be of interest to the readers to learn a little bit more about the differences/similarities between Beiwe and other platforms/software.
