### 2021-06-30 onnela-lab/beiwe-backend#210 @erik-whiting
Peer Review - Documentation Issues
Some issues from my review of the documentation. Review is being coordinated [here](https://github.com/openjournals/joss-reviews/issues/3417)

**_Installation Instructions_** - There are several issues with the development setup:
  * The line in step one `sudo apt-get install postgresql python-psycopg2 libpq-dev` breaks because `python-psycopg2` is a package you can install with pip. The package is in your `requirements.txt` file, so you don't need special instructions to install it.
  * The last line of your suggested `private/environment.sh` file has the sysadmin email as "sysadmin@localhst" is this a typo or is that the correct email address?
  * Installing the requirements from `requirements.txt` failed with message `error: launchpadlib 1.10.13 requires testresources`. I ran `pip3 install testresources` which fixed the error but you should include that package in `requirements.txt` since it isn't a standard library module
  * Under "Local Celery setup" the link provided in step 1 is broken and goes to a 404 page. I had to install rabbit mq manually
  * The line `sudo rabbitmqctl set_permissions -p / beiwe ".*" ".*" ".*"` failed because there's no such user "beiwe." Am I supposed to use my own user or was there a step I missed / step that wasn't documented?
  * Running celery after this line doesn't work because the connection is refused. I'm guessing it has something to do with the above issue
  * Step 7 doesn't work either, I assume because of the two issues above, but it's also throwing database errors. I see that there is a Django app in this repo, are there some Django setup steps missing from the README?
  * There is a wiki for running locally located here: https://github.com/onnela-lab/beiwe-backend/wiki/How-to-Run-Beiwe-Locally, but it is outdated compared to when the README was last updated, so I'm not sure which directions to follow

**_Example Usage_** - The checklist item for this says "Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)" I do not see such examples. Is this solely the backend for the Beiwe mobile app? i.e., are there not really any examples to show here? If so, I guess no examples are needed.

**_Some Concerns_** - Not listed on the checklist but I have a concern. In the README, you say this system is built with Python 3.6. Python 3.6 end-of-life is December 23rd, 2021. That's about six months from now. If you were to add a little more test coverage, you could set up a build server like CircleCI for free and run the test suite with multiple Python versions pretty easily and be able to deal with EOLs confidently. If this is something you're interested in pursuing and would like help with, I don't mind pointing you in the right direction or helping out a bit. If you _**do**_ have CI/CD set up already, it is not obvious, but I apologize for missing it if I did.

I cannot review the areas in the Functionality portion of the checklist until the installation issues are fixed or a workaround is suggested
