### 2021-06-30 #3417 @erik-whiting
I opened [this](https://github.com/onnela-lab/beiwe-backend/issues/210) issue as my review to their installation.

I cannot review the `functionality` portion of the checklist until the installation issues are fixed or someone offers a workaround. Otherwise, I'm done with my first round of reviewing.
