### 2021-06-30 onnela-lab/beiwe-backend#210 @biblicabeebli
Hello! I've been following your posts, this is very interesting and quite helpful!  You've caught me (us, really, though I don't want to speak for others) at an odd moment.  I would love to get in contact, and be a little more active in responding and commenting as soon as time and commitments allow.

Would greatly appreciate it if you could send me an email so we could get in touch, I've set up biblicabeiwe at gmail dot com.  (let me know if that isn't working.)
