+++ database/migrations/0059_devicesettings_ambient_audio.py
@@ -1,4 +1,4 @@
-# Generated by Django 2.2.14 on 2021-04-06 03:43
+# Generated by Django 2.2.19 on 2021-07-06 19:58
 
 from django.db import migrations, models
 
@@ -6,7 +6,7 @@ from django.db import migrations, models
 class Migration(migrations.Migration):
 
     dependencies = [
-        ('database', '0050_participant_unregistered'),
+        ('database', '0058_auto_20210701_2222'),
     ]
 
     operations = [
