### 2021-07-09 #3417 @erik-whiting
>  @erik-whiting, please update us on how your review is going (this is an automated reminder).

I opened two issues, one for documentation and one for the paper. I cannot review the functionality section until the documentation issues are addressed
