### 2021-07-10 #3417 @Tam-Pham
> @Tam-Pham, please update us on how your review is going (this is an automated reminder).

I'm currently following the issue opened by @erik-whiting regarding the development setup (https://github.com/onnela-lab/beiwe-backend/issues/210). I too ran into the same problem that @erik-whiting faced. Erik has cleverly summarized those in the issue that was opened and hence I will wait till those are addressed before continuing with my review.
