### 2021-08-17 1871750
[firebase-cleanup] separates firebase configuration from push notification logic (import fixes on rebase)
