### 2021-09-01 onnela-lab/beiwe-backend#226 @biblicabeebli
Review and merge all documentation for the iOS, Android, and Backend into a single wiki.
(And, in my humble opinion, the data science work on Forest too, but that's a larger conversation.)

I am unable to find documentation that I am positive exists, frequently because I remember writing it.  This is irksome and a problem, I can't effectively direct people to the appropriate documentation.  We _do_ have some decent SEO via the longer bug reports on github, this is intentional and effective, but it could be better.

The original researcher's guide, written by the original project manager starting in early development is MIA.
