### 2021-09-01 0fe800c
[fix-deployment-for-password-protected-ssh-keys] Sets fabric to not explode on prompts. While blocking prompts entirely is useful in development, AWS can provide keys with passwords - I seem to recall this was not a thing when originally developed - and we have had some issues where this was the culprit.
