+++ cluster_management/launch_script.py
@@ -55,7 +55,7 @@ from deployment_helpers.general_utils import current_time_string, do_zip_reducti
 # Fabric configuration
 class FabricExecutionError(Exception): pass
 fabric_env.abort_exception = FabricExecutionError
-fabric_env.abort_on_prompts = True
+fabric_env.abort_on_prompts = False
 
 parser = argparse.ArgumentParser(description="interactive set of commands for deploying a Beiwe Cluster")
 
@@ -110,8 +110,8 @@ def push_manager_private_ip_and_password(eb_environment_name):
     # echo puts a new line at the end of the output
     run(f"echo {ip} > {REMOTE_RABBIT_MQ_PASSWORD_FILE_PATH}")
     run(f"printf {password} >> {REMOTE_RABBIT_MQ_PASSWORD_FILE_PATH}")
-    
-    
+
+
 def push_home_directory_files():
     for local_relative_file, remote_relative_file in FILES_TO_PUSH:
         local_file_path = path_join(PUSHED_FILES_FOLDER, local_relative_file)
@@ -192,7 +192,7 @@ def setup_rabbitmq(eb_environment_name):
     # push the configuration file so that it listens on the configured port
     put(LOCAL_RABBIT_MQ_CONFIG_FILE_PATH, REMOTE_RABBIT_MQ_CONFIG_FILE_PATH)
     sudo(f"cp {REMOTE_RABBIT_MQ_CONFIG_FILE_PATH} {REMOTE_RABBIT_MQ_FINAL_CONFIG_FILE_PATH}")
-    
+
     # setup a new password
     sudo(f"rabbitmqctl add_user beiwe {get_rabbit_mq_password(eb_environment_name)}")
     sudo('rabbitmqctl set_permissions -p / beiwe ".*" ".*" ".*"')
@@ -295,7 +295,7 @@ def do_fail_if_bad_environment_name(name):
     if not (4 <= len(name) < 40):
         log.error("That name is either too long or too short.")
         EXIT(1)
-    
+
     if not re.match("^[a-zA-Z0-9-]+$", name) or name.endswith("-"):
         log.error("that is not a valid Elastic Beanstalk environment name.")
         EXIT(1)
@@ -326,13 +326,13 @@ def prompt_for_extant_eb_environment_name():
 
 def do_setup_eb_update():
     print("\n", DO_SETUP_EB_UPDATE_OPEN)
-    
+
     files = sorted([f for f in os.listdir(STAGED_FILES) if f.lower().endswith(".zip")])
-    
+
     if not files:
         print("Could not find any zip files in " + STAGED_FILES)
         EXIT(1)
-    
+
     print("Enter the version of the codebase do you want to use:")
     for i, file_name in enumerate(files):
         print("[%s]: %s" % (i + 1, file_name))
@@ -343,11 +343,11 @@ def do_setup_eb_update():
         log.error("Could not parse input.")
         index = None  # ide warnings
         EXIT(1)
-    
+
     if index < 1 or index > len(files):
         log.error("%s was not a valid option." % index)
         EXIT(1)
-    
+
     # handle 1-indexing
     file_name = files[index - 1]
     # log.info("Processing %s..." % file_name)
@@ -381,34 +381,34 @@ def do_help_setup_new_environment():
     beiwe_environment_fp = get_beiwe_python_environment_variables_file_path(name)
     processing_server_settings_fp = get_server_configuration_file_path(name)
     extant_files = os.listdir(DEPLOYMENT_SPECIFIC_CONFIG_FOLDER)
-    
+
     for fp in (beiwe_environment_fp, processing_server_settings_fp):
         if os.path.basename(fp) in extant_files:
[…]
