### 2021-09-06 757e6ce
Performance improvements on View Study page. Load participants table via AJAX as needed, rather than all at once on pageload.
Minor text change, rename modal

DataTables use server-side pagination- initial

Load DataTables JS & CSS from CDN, also only on View Study page

Fix draw param on DataTables AJAX, now reloads properly

Add sorting and searching

Get phone registered as boolean. Format created_on date.

Enable searching on OS type

Change view participant URL to use 8-digit patient ID, not PK

Correctly report total # of participants and filtered # of participants

Make participant ID a
[…]
