### 2021-09-10 #3417 @fboehm
I sent the following email message to the submitting author's email address today:

Hi, Professor Onnela,

I'm writing as a JOSS editor to update you on the status of your submission (Beiwe: A data collection platform for high-throughput digital phenotyping) to Journal of Open Source Software.

It seems that there have been no responses from you and your team in more than a month. Please begin to address the issues raised by the reviewers and indicate progress on them before September 30, 2021.

You and your team should feel free to discuss the issues raised by reviewers in the review thread: https://github.com/openjournals/joss-reviews/issues/3417

If this timeline seems unreasonable, please indicate what would be a reasonable timeline.

Thank you,
fred
