### 2021-09-10 #3417 @fboehm
I followed up with this email message:

Professor Onnela,
I need to amend the earlier statement to say that, if no progress is noted by September 30, we'll need to reject the submission.
Thanks for your attention to this.
fred
