### 2021-09-11 ff22fb8
Remove an unnecessary line
