### 2021-09-11 a38c78a
Add type annotations
