### 2021-09-14 529b92a
Move participants data table query from user_models to study_models
