### 2021-09-15 7f68c1e
Use prefetch_related to cut DB calls on participants table
