### 2021-09-15 #3417 @fboehm
Today I sent this email to Eli Jones, a collaborator of the submitting author. I copied the submitting author, too.

Hi, Eli,

Thanks so much! I really appreciate your enthusiasm and look forward to guiding you through the review process.

The only deadline is for, at a minimum, demonstration of further discussion - on the Github issue https://github.com/openjournals/joss-reviews/issues/3417 and on the issues that the reviewers have opened on your repository. If this process can begin as soon as possible, and no later than Sept 30, then we can continue with the review and ultimately publish the submission.

Ideally, we'd like to see the review completed within about 6 to 8 weeks. We do have some flexibility on the duration of the review so that, as long as continuous progress is made, there won't be additional deadlines that could lead to rejection of the submission. The only reason for imposing the Sept 30 deadline was the apparent inactivity on the review and a desire to not unnecessarily prolong the involvement of our reviewers.

Additionally, I encourage both you and our reviewers to avoid email communications in favor of posting communications on the "review" issue on Github. This way, we can have a more transparent record of what is said. We'll also then have a record that's more accessible. In keeping with that spirit, I'll post the text of this email to the review thread right away.

Thanks again!

fred

PS - please feel free to ask me any questions along the way. My github username is fboehm, so mentioning me with an at sign (eg, @fboehm) in the issue will get my attention.
