### 2021-09-15 8f58524
[development] updates to push notification logic, comments, adding a value and documenting a couple things better in settings
