+++ config/settings.py
@@ -1,11 +1,11 @@
 from os import cpu_count, getenv
 
 """
-To customize any of these values, append a line to config/remote_db_env.py such as:
-os.environ['S3_BUCKET'] = 'bucket_name'
+On processing servers, instead of environment varables, append a line to config/remote_db_env.py,
+such as: os.environ['S3_BUCKET'] = 'bucket_name'
 """
 
-
+# for we retain the old access key in case someone deployed a super old EB server.
 BEIWE_SERVER_AWS_ACCESS_KEY_ID = getenv("BEIWE_SERVER_AWS_ACCESS_KEY_ID") or getenv("S3_ACCESS_CREDENTIALS_USER")
 BEIWE_SERVER_AWS_SECRET_ACCESS_KEY = getenv("BEIWE_SERVER_AWS_SECRET_ACCESS_KEY") or getenv("S3_ACCESS_CREDENTIALS_KEY")
 
@@ -13,26 +13,27 @@ BEIWE_SERVER_AWS_SECRET_ACCESS_KEY = getenv("BEIWE_SERVER_AWS_SECRET_ACCESS_KEY"
 #  cryptographically secure string to this value.
 FLASK_SECRET_KEY = getenv("FLASK_SECRET_KEY")
 
-# the name of the s3 bucket that will be used to store user generated data, and backups of local
+# The name of the s3 bucket that will be used to store user generated data, and backups of local
 # database information.
 S3_BUCKET = getenv("S3_BUCKET")
 
 # Domain name for the server
 DOMAIN_NAME = getenv("DOMAIN_NAME")
 
-# A list of email addresses that will receive error emails. This value must be a
-# comma separated list; whitespace before and after addresses will be stripped.
+# A list of email addresses that will receive error emails. This value must be a comma separated
+# list; whitespace before and after addresses will be stripped.
 SYSADMIN_EMAILS = getenv("SYSADMIN_EMAILS")
 
-# Sentry DSNs (optional)
+# Sentry DSNs (technically optional)
 SENTRY_DATA_PROCESSING_DSN = getenv("SENTRY_DATA_PROCESSING_DSN")
 SENTRY_ELASTIC_BEANSTALK_DSN = getenv("SENTRY_ELASTIC_BEANSTALK_DSN")
 SENTRY_JAVASCRIPT_DSN = getenv("SENTRY_JAVASCRIPT_DSN")
 
-# Production/Staging: set to "TRUE" if staging
+# Production/Staging: set to "TRUE" if this is a staging/testing/development server to enable some
+# extra features.
 IS_STAGING = getenv("IS_STAGING") or "PRODUCTION"
-REPORT_DECRYPTION_KEY_ERRORS = getenv("REPORT_DECRYPTION_KEY_ERRORS") or False
-STORE_DECRYPTION_KEY_ERRORS = getenv("STORE_DECRYPTION_KEY_ERRORS") or False
+REPORT_DECRYPTION_KEY_ERRORS = bool(getenv("REPORT_DECRYPTION_KEY_ERRORS", False))
+STORE_DECRYPTION_KEY_ERRORS = bool(getenv("STORE_DECRYPTION_KEY_ERRORS", False))
 
 # S3 region (not all regions have S3, so this value may need to be specified)
 S3_REGION_NAME = getenv("S3_REGION_NAME", "us-east-1")
@@ -43,11 +44,16 @@ DOWNLOADABLE_APK_URL = getenv("DOWNLOADABLE_APK_URL", "https://s3.amazonaws.com/
 # File processing directives
 # Used in data download and data processing, base this on CPU core count.
 CONCURRENT_NETWORK_OPS = getenv("CONCURRENT_NETWORK_OPS") or cpu_count() * 2
-# Used in file processing, number of files to be pulled in and processed simultaneously.
[…]
