### 2021-09-16 #3417 @biblicabeebli
(waaaooow I left this unsubmitted in an open tab.  punching the comment button now for completeness I guess.)

Hello!
I'm working with Professor Onnela, I'm one of the primary maintainers and original developers of the Beiwe platform.  I just joined the lab as an employee and can be available for this immediately.  I have an email thread open with Erik, feel free to get in direct contact with me.  I'm working on some delayed bug fixes, we have a github issue tracking Erik's requested documentation updates, and I hope to get to that soon.
