### 2021-09-17 8ede542
[fix-copy-study-relative-schedules] many changes so that we can export relative schedules and their interventions. Contains code to maintain forwards compatibility as best as possible with exports from old versions of beiwe backend.
