+++ api/copy_study_api.py
@@ -3,10 +3,12 @@ from json import dumps, loads
 from flask import abort, Blueprint, flash, redirect, request, Response
 
 from authentication.admin_authentication import authenticate_admin
+from database.schedule_models import Intervention
 from database.study_models import Study
 from libs.copy_study import (ABSOLUTE_SCHEDULE_KEY, add_new_surveys, allowed_file_extension,
-    DEVICE_SETTINGS_KEY, purge_unnecessary_fields, RELATIVE_SCHEDULE_KEY, SURVEYS_KEY,
-    update_device_settings, WEEKLY_SCHEDULE_KEY)
+    DEVICE_SETTINGS_KEY, INTERVENTIONS_KEY, purge_unnecessary_fields, RELATIVE_SCHEDULE_KEY,
+    SURVEYS_KEY, update_device_settings, WEEKLY_SCHEDULE_KEY)
+
 
 copy_study_api = Blueprint('copy_study_api', __name__)
 
@@ -32,10 +34,9 @@ def export_study_settings_file(study_id):
         # content, cleanup, then schedules.
         survey_content = survey.as_unpacked_native_python()
         purge_unnecessary_fields(survey_content)
-
         survey_content[WEEKLY_SCHEDULE_KEY] = survey.weekly_timings()
         survey_content[ABSOLUTE_SCHEDULE_KEY] = survey.absolute_timings()
-        survey_content[RELATIVE_SCHEDULE_KEY] = survey.relative_timings()
+        survey_content[RELATIVE_SCHEDULE_KEY] = survey.relative_timings_by_name()
         surveys.append(survey_content)
 
     device_settings = study.device_settings.as_unpacked_native_python()
@@ -44,6 +45,7 @@ def export_study_settings_file(study_id):
     output = {
         SURVEYS_KEY: surveys,
         DEVICE_SETTINGS_KEY: device_settings,
+        INTERVENTIONS_KEY: list(study.interventions.value_list("name", flat=True))
     }
 
     filename = study.name.replace(' ', '_') + "_surveys_and_settings.json"
@@ -70,9 +72,18 @@ def import_study_settings_file(study_id):
     study_data = loads(file.read())
     device_settings = study_data.pop(DEVICE_SETTINGS_KEY, None)
     surveys = study_data.pop(SURVEYS_KEY, None)
+    interventions = study_data.pop(INTERVENTIONS_KEY, None)
+
+    # create all interventions in the json that don't already exist.
+    if interventions and isinstance(interventions, list):
+        extant_interventions = set(study.interventions.values_list("name", flat=True))
+        Intervention.objects.bulk_create(
+            [Intervention(name=name, study=study)
+             for name in interventions if name not in extant_interventions]
+        )
 
     # these functions return a message to construct for the user
     msg = update_device_settings(device_settings, study, file.filename)
-    msg += " \n" + add_new_surveys(surveys, study, file.filename)
+    msg += " \n" + add_new_surveys(surveys, study, file.filename, from_json=True)
     flash(msg, 'success')
     return redirect('/edit_study/{:s}'.format(study_id))
