### 2021-09-17 ae170c8
[fix-copy-study-relative-schedules] turns out there aren't usually enough interventions for all that effort to be worth it.
