### 2021-09-21 b31bea2
Start of a method to get data qty from ChunkRegistry
