### 2021-09-22 2be55c9
Initial code to save stats after file processing updates chunks
