### 2021-09-23 3deff8b
Working code to save DataQuantities, with a few data streams
