+++ libs/file_processing/data_qty_stats.py
@@ -1,26 +1,29 @@
-from datetime import datetime, timedelta
+from collections import defaultdict
 from dateutil import tz
-from django.db import models
 
 from database.data_access_models import ChunkRegistry
+from database.tableau_api_models import DataQuantity
+from database.user_models import Participant
+
+ALLOWED_DATA_STREAMS = {'accelerometer', 'app_log'}  # TODO: add more. Maybe also make this on the DataQuantity model.
 
 
-def calculate_data_quantity_stats(participant_id: int):
-    earliest_chunk_registry = (ChunkRegistry.objects
-                               .filter(participant_id=participant_id).earliest('time_bin'))
-    print(earliest_chunk_registry.time_bin)  # Datetime
-
-
-def get_data_qty(participant_id: int, day: datetime.date, study_timezone: datetime.tzinfo):
-    """ Return the data quantity, in bytes, for a specific participant, data stream, and day. """
-    start_time_local = datetime(day.year, day.month, day.day, tzinfo=study_timezone)
-    end_time_local = start_time_local + timedelta(days=1)
-    start_time = start_time_local.astimezone(tz.UTC)
-    end_time = end_time_local.astimezone(tz.UTC)
-    return (ChunkRegistry.objects
-            .filter(participant_id=participant_id, time_bin__gte=start_time, time_bin__lt=end_time)
-            .aggregate(models.Sum('file_size')))
-
-
-# Participant.objects.annotate(totalbytes=Sum('chunk_registries__file_size')).values_list('patient_id', 'totalbytes').order_by('-totalbytes')
-
+def calculate_data_quantity_stats(participant: Participant, study_timezone: tz.tzfile):
+    """ Update the daily DataQuantity stats for a participant """
+    chunkregistries = ChunkRegistry.objects.filter(participant=participant).all()
+    daily_data_qtys = defaultdict(lambda: defaultdict(lambda: 0))
+    for chunkregistry in chunkregistries:
+        day = chunkregistry.time_bin.astimezone(study_timezone).date()
+        daily_data_qtys[day][chunkregistry.data_type] += chunkregistry.file_size
+    for day in daily_data_qtys:
+        data_qty = {
+            "participant": participant,
+            "date": day,
+        }
+        for data_type in daily_data_qtys[day]:
+            if data_type in ALLOWED_DATA_STREAMS:
+                data_qty[data_type + '_bytes'] = daily_data_qtys[day][data_type]
+        print(data_qty)
+        DataQuantity.objects.update_or_create(**data_qty)
+        # TODO: bulk update or create
+        # TODO: use objects.exclude to delete those objects that shouldn't exist any longer
