### 2021-09-23 1b92965
Add all data types. Alphabetize some of them in constants.py
