+++ database/migrations/0060_auto_20210923_0516.py
@@ -1,0 +1,48 @@
+# Generated by Django 2.2.24 on 2021-09-23 05:16
+
+from django.db import migrations, models
+import django.db.models.deletion
+
+
+class Migration(migrations.Migration):
+
+    dependencies = [
+        ('database', '0059_devicesettings_ambient_audio'),
+    ]
+
+    operations = [
+        migrations.CreateModel(
+            name='DataQuantity',
+            fields=[
+                ('id', models.AutoField(auto_created=True, primary_key=True, serialize=False, verbose_name='ID')),
+                ('created_on', models.DateTimeField(auto_now_add=True)),
+                ('last_updated', models.DateTimeField(auto_now=True)),
+                ('date', models.DateField(db_index=True)),
+                ('accelerometer_bytes', models.IntegerField(blank=True, null=True)),
+                ('ambient_audio_bytes', models.IntegerField(blank=True, null=True)),
+                ('app_log_bytes', models.IntegerField(blank=True, null=True)),
+                ('bluetooth_bytes', models.IntegerField(blank=True, null=True)),
+                ('calls_bytes', models.IntegerField(blank=True, null=True)),
+                ('devicemotion_bytes', models.IntegerField(blank=True, null=True)),
+                ('gps_bytes', models.IntegerField(blank=True, null=True)),
+                ('gyro_bytes', models.IntegerField(blank=True, null=True)),
+                ('identifiers_bytes', models.IntegerField(blank=True, null=True)),
+                ('image_survey_bytes', models.IntegerField(blank=True, null=True)),
+                ('ios_log_bytes', models.IntegerField(blank=True, null=True)),
+                ('magnetometer_bytes', models.IntegerField(blank=True, null=True)),
+                ('power_state_bytes', models.IntegerField(blank=True, null=True)),
+                ('proximity_bytes', models.IntegerField(blank=True, null=True)),
+                ('reachability_bytes', models.IntegerField(blank=True, null=True)),
+                ('survey_answers_bytes', models.IntegerField(blank=True, null=True)),
+                ('survey_timings_bytes', models.IntegerField(blank=True, null=True)),
+                ('texts_bytes', models.IntegerField(blank=True, null=True)),
+                ('audio_recordings_bytes', models.IntegerField(blank=True, null=True)),
+                ('wifi_bytes', models.IntegerField(blank=True, null=True)),
+                ('participant', models.ForeignKey(on_delete=django.db.models.deletion.CASCADE, related_name='data_quantities', to='database.Participant')),
+            ],
+        ),
+        migrations.AddConstraint(
+            model_name='dataquantity',
+            constraint=models.UniqueConstraint(fields=('date', 'participant'), name='unique_data_quantity'),
+        ),
+    ]
