+++ database/tableau_api_models.py
@@ -346,8 +346,25 @@ class DataQuantity(TimestampedModel):
     date = models.DateField(db_index=True)
 
     accelerometer_bytes = models.IntegerField(null=True, blank=True)
+    ambient_audio_bytes = models.IntegerField(null=True, blank=True)
     app_log_bytes = models.IntegerField(null=True, blank=True)
-    # TODO: add more data streams
+    bluetooth_bytes = models.IntegerField(null=True, blank=True)
+    calls_bytes = models.IntegerField(null=True, blank=True)
+    devicemotion_bytes = models.IntegerField(null=True, blank=True)
+    gps_bytes = models.IntegerField(null=True, blank=True)
+    gyro_bytes = models.IntegerField(null=True, blank=True)
+    identifiers_bytes = models.IntegerField(null=True, blank=True)
+    image_survey_bytes = models.IntegerField(null=True, blank=True)
+    ios_log_bytes = models.IntegerField(null=True, blank=True)
+    magnetometer_bytes = models.IntegerField(null=True, blank=True)
+    power_state_bytes = models.IntegerField(null=True, blank=True)
+    proximity_bytes = models.IntegerField(null=True, blank=True)
+    reachability_bytes = models.IntegerField(null=True, blank=True)
+    survey_answers_bytes = models.IntegerField(null=True, blank=True)
+    survey_timings_bytes = models.IntegerField(null=True, blank=True)
+    texts_bytes = models.IntegerField(null=True, blank=True)
+    audio_recordings_bytes = models.IntegerField(null=True, blank=True)
+    wifi_bytes = models.IntegerField(null=True, blank=True)
 
     class Meta:
         constraints = [
