### 2021-09-23 1019e79
Delete DataQuantity objects that shouldn't exist
