### 2021-09-23 5820e65
Bulk create; simplify code
