+++ libs/file_processing/data_qty_stats.py
@@ -1,5 +1,4 @@
 from collections import defaultdict
-from django.db.models import Q
 
 from config.constants import ALL_DATA_STREAMS
 from database.data_access_models import ChunkRegistry
@@ -16,7 +15,10 @@ def calculate_data_quantity_stats(participant: Participant):
     for chunkregistry in chunkregistries:
         day = chunkregistry.time_bin.astimezone(study_timezone).date()
         daily_data_qtys[day][chunkregistry.data_type] += chunkregistry.file_size
+    # Delete all existing DataQuantity objects for the participant
+    DataQuantity.objects.filter(participant=participant).delete()
     # For each date, create a DataQuantity object
+    data_qty_objects_to_be_created = []
     for day in daily_data_qtys:
         data_qty = {
             "participant": participant,
@@ -25,10 +27,8 @@ def calculate_data_quantity_stats(participant: Participant):
         for data_type in daily_data_qtys[day]:
             if data_type in ALL_DATA_STREAMS:
                 data_qty[data_type + '_bytes'] = daily_data_qtys[day][data_type]
-        DataQuantity.objects.update_or_create(**data_qty)
-    # Delete any DataQuantity objects that shouldn't exist, i.e. from days with no ChunkRegistries
-    all_days_with_data_query = Q()
-    for day in daily_data_qtys:
-        all_days_with_data_query |= Q(date=day)
-    DataQuantity.objects.filter(participant=participant).exclude(all_days_with_data_query).delete()
-    # TODO: bulk update or create
+        data_qty_objects_to_be_created.append(DataQuantity(**data_qty))
+    DataQuantity.objects.bulk_create(data_qty_objects_to_be_created)
+    # TODO: if this needs to be optimized in the future, we could improve performance by only
+    # querying ChunkRegistry objects that are newer than the last time we calculated all the
+    # DataQuantity objects. But that would make the code significantly more complicated.
