### 2021-09-23 #3417 @biblicabeebli
Not at the moment!  I have been stitching some important bug fixes together, just starting to update github issues, post new ones.

Your two posts are https://github.com/onnela-lab/beiwe-backend/issues/210 and https://github.com/onnela-lab/beiwe-backend/issues/208.

The documentation issues are-in-fact helpful, and we have [another issue tracking merging in the old wiki](https://github.com/onnela-lab/beiwe-backend/issues/226) that necessitates a more generic review of content (including for correctness).

Any asks from me:
... None... yet...
