### 2021-09-24 457190f
Use .values() to make the query more lightweight
