### 2021-09-24 04bd3e6
Merge DataQuantity into existing SummaryStatsDaily model
