### 2021-09-25 614e5e8
Move calculate_data_qty_stats() to where it should always be called whenever ChunkRegistries are updated
