### 2021-09-27 onnela-lab/beiwe-backend#226 @biblicabeebli
And now there is a landing page.
I don't have edit access to the old wiki, not closing this issue until I can update all the pages over there that that wiki is deprecated.
