### 2021-09-27 onnela-lab/beiwe-backend#210 @biblicabeebli
- The line in step one sudo apt-get install postgresql python-psycopg2 libpq-dev breaks because python-psycopg2 is a package you can install with pip. The package is in your requirements.txt file, so you don't need special instructions to install it.

I've addressed this documentation on the branch `peer-review`.  there's more to get through so no pull request yet.  I believe the origin of this detail was related to virtual environments (or a lack thereof), so I added a statement about virtual environments beforehand.
