### 2021-09-27 onnela-lab/beiwe-backend#210 @biblicabeebli
- The last line of your suggested private/environment.sh file has the sysadmin email as "sysadmin@localhst" is this a typo or is that the correct email address?

lol, typo. pushed on `peer-review`
