### 2021-09-27 onnela-lab/beiwe-backend#210 @biblicabeebli
- Installing the requirements from requirements.txt failed with message error: launchpadlib 1.10.13 requires testresources. I ran pip3 install testresources which fixed the error but you should include that package in requirements.txt since it isn't a standard library module

I don't know what you encountered here, launchpadlib is not a dependency of any library in use right now*, and I don't think it was at the time.  More relevantly, pip should have installed that on the first go?  Maybe you were running an older version of pip and some bug resulted in a missing dependency?  I don't know what's going on here, and I can't replicate.

That said, I will go ahead and add a `pip install --upgrade pip setuptools wheel` to the instructions to avoid pip-related issues.
