### 2021-09-27 onnela-lab/beiwe-backend#210 @biblicabeebli
- Under "Local Celery setup" the link provided in step 1 is broken and goes to a 404 page. I had to install rabbit mq manually

fixed on peer-review
