### 2021-09-27 onnela-lab/beiwe-backend#210 @erik-whiting
> * Installing the requirements from requirements.txt failed with message error: launchpadlib 1.10.13 requires testresources. I ran pip3 install testresources which fixed the error but you should include that package in requirements.txt since it isn't a standard library module
>
> I don't know what you encountered here, launchpadlib is not a dependency of any library in use right now, and I don't think it was at the time. More relevantly, pip should have installed that on the first go? Maybe you were running an older version of pip and some bug resulted in a missing dependency? I don't know what's going on here, and I can't replicate.
>
> That said, I will go ahead and add a `pip install --upgrade pip setuptools wheel` to the instructions to avoid pip-related issues.
>
> edit: had a random * hanging out in the text...

@biblicabeebli the line you added still throws an error for me. I was able to get around it by running `sudo apt install python3-testresources`. I'm not really sure what was causing the error and maybe it's something on my side only (I am blowing away everything before running the setup tasks to simulate a brand-new contributor with a brand-new machine trying to get started with this repo).

I defer to you if you want to add that line to your setup script or not since I'm not sure if it's a me-specific problem; if you choose to leave it out of the setup script I'll still mark this part of the issue as fixed.
