### 2021-09-27 onnela-lab/beiwe-backend#210 @biblicabeebli
- Under "Local Celery setup" the link provided in step 1 is broken and goes to a 404 page. I had to install rabbit mq manually
- The line sudo rabbitmqctl set_permissions -p / beiwe ".*" ".*" ".*" failed because there's no such user "beiwe." Am I supposed to use my own user or was there a step I missed / step that wasn't documented?
- Running celery after this line doesn't work because the connection is refused. I'm guessing it has something to do with the above issue

I've gone and updated this section considerably.  `rabbitmq` is such a pain to work with that I ended up writing ~dirty~ clever a hack that bypasses it entirely for local testing purposes.  We only bother with Celery+rabbitmq on the data processing servers now, which have automated deployment and configuration.  Directions are as good and complete as I can make them, last I checked Ubuntu 20.04+ was magically incompatible with those directions, and I've noted that.
