### 2021-09-27 onnela-lab/beiwe-backend#210 @erik-whiting
> > @biblicabeebli the line you added still throws an error for me. I was able to get around it by running `sudo apt install python3-testresources`. I'm not really sure what was causing the error and maybe it's something on my side only (I am blowing away everything before running the setup tasks to simulate a brand-new contributor with a brand-new machine trying to get started with this repo).
>
> Hunh. Um. Are you using a virtual environment, or the system environment?

virtual
