### 2021-09-27 onnela-lab/beiwe-backend#210 @biblicabeebli
python3-testresources is another pip install masquerading as a system install for the system python environment.  Maybe it is installing some other, related system dependency that you are missing? buh?

I'm using this to get legible output from apt, and I've followed a couple dependencies, it doesn't really go that deep... python3, six, python3 packaging junk...
`apt show python3-testresources| grep ^Depends`

I guess our instructions don't actually state the other more common system dependencies, we only declare the third-party binaries that you may be missing and wouldn't just-incidentally-know without interrogating the python requirements file and already knowing voodoo facts about python. (Like "oh yeah, psycopg2 is the postgres library because that's totally obvious... oh and it has a compilation step as part of a python install... unless you declare it as 'psycopg2-binary'...")

I wouldn't and didn't think to include things like gcc (really any compiler) as a required dependency, that just seemed to basic.  This isn't a line of thought I've reflected on much.  I have no problem adding a couple extra items like python3 and gcc to the apt installs, it costs us nothing and could help someone out eventually, I'm just not sure where that line is.... but I do know that it is not this obscure testing library.

Could you post the command and output that is stating it is missing this plus the output of pip list?
