### 2021-09-28 3c3dc7c
Only calculate data qty stats for time range of updated ChunkRegistries Instead of all time
