### 2021-09-28 onnela-lab/beiwe-backend#210 @biblicabeebli
https://github.com/onnela-lab/beiwe-backend/wiki/Tips-For-Local-Beiwe-Development

- There is a wiki for running locally located here: https://github.com/onnela-lab/beiwe-backend/wiki/How-to-Run-Beiwe-Locally, but it is outdated compared to when the README was last updated, so I'm not sure which directions to follow

Yup, that was confusing.  I've updated the name to more accurately reflect what the page is for, and I'm adding that link to the readme.

new link [https://github.com/onnela-lab/beiwe-backend/wiki/Tips-For-Local-Beiwe-Development](https://github.com/onnela-lab/beiwe-backend/wiki/Tips-For-Local-Beiwe-Development)
