### 2021-09-28 onnela-lab/beiwe-backend#210 @biblicabeebli
- Some Concerns - Not listed on the checklist but I have a concern. In the README, you say this system is built with Python 3.6. Python 3.6 end-of-life is December 23rd, 2021. That's about six months from now. If you were to add a little more test coverage, you could set up a build server like CircleCI for free and run the test suite with multiple Python versions pretty easily and be able to deal with EOLs confidently. If this is something you're interested in pursuing and would like help with, I don't mind pointing you in the right direction or helping out a bit. If you do have CI/CD set up already, it is not obvious, but I apologize for missing it if I did.

Yup.  This work got delayed due to personnel (me) issues in moving to be more directly associated with Onnela Lab.  Annoyingly the AWS Python 3.8 environment for Elastic Beanstalk (automated scaling for server deployment) swaps some technology on us, primarily Apache -> nginx.  I maintain the deployment code, but at the moment I'm catching up on bugs, this JOSS review work, and some long-overdue tech issues that need addressing and will ideally be out ~along side the 3.8 (3.7? arg, really need to read up) updates.

Fortunately, I have an experienced colleague who's done this particular major elastic beanstalk version migration, and we already have a substantially more complex one under our belts for the python 2.7-3.6 transition.  Unlike python 2, there are zero code compatibility issues at play.  I've got a local pyenv environment congured on a newer version of python that I can and have swapped to on occasion.

(I even have a branch, `eli/type-annotations-python-37`, that includes a nice type annotation feature (its a __future__ import introduced in 3.7) that allows class functions can declare return type annotations, an otherwise class-static scope, with the type of the class they are contained within.  Super helpful in the IDE.)
