### 2021-09-28 onnela-lab/beiwe-backend#210 @biblicabeebli
-  I see that there is a Django app in this repo, are there some Django setup steps missing from the README?

Among the tech upgrades are swapping out Flask for a full django tech stack.  The reason the code is structured as it is is entirely historical.  This started as a flask-mongo project - what can I say, mistakes were made - and we transitioned from mongo to django+postgres a couple years ago now, but the hack to get flask + django working is... a hack.  The work required to replace all of flask with django frontend is daunting, but I was able to start on that recently and have the work up on the eli/django branch.

Once the Django code is done the amount of overhead required to have a functional test suite will be _substantially_ reduced, and oh-lordy will I be writing a lot of tests.  Then comes a CI.  And probably then a vacation.
