### 2021-09-28 onnela-lab/beiwe-backend#210 @biblicabeebli
- Example Usage - The checklist item for this says "Do the authors include examples of how to use the software (ideally to solve real-world analysis problems)" I do not see such examples. Is this solely the backend for the Beiwe mobile app? i.e., are there not really any examples to show here? If so, I guess no examples are needed.

First - is this the checklist over on https://github.com/openjournals/joss-reviews/issues/3417 ?  I have not _really_ familiarized myself much about JOSS, I need some butt-in-chair time to go through the checklist and work out the questions I / Beiwe needs to ask.

I suspect the answer to your point above is that 1) I'm on the developer side, not the data analysis side, so we may need to pull someone else in, and 2) the locus of the analysis work for Beiwe is currently within the Forest repo, which is here: [https://github.com/onnela-lab/forest](https://github.com/onnela-lab/forest).  Historically there have been other relevant repositories, but they are, as far as I know, totally deprecated.
