### 2021-09-28 onnela-lab/beiwe-backend#210 @biblicabeebli
I believe I have addressed everything in the original post here.  I'm getting notifications on this thread.  Let me know if there is something I have missed.

Potential relevant note: I'm going to be reviewing the (newly copied in) researcher's section of the backend wiki, which is known to be outdated and incorrect.  The tl;dr there is that we lost track of it because it was attached to an old repo.  Literally, we forgot about it.  This is being tracked on issue #226. Further, there are details on that wiki section that are related to data science aspects that I'm simply not an expert in.
