### 2021-09-29 onnela-lab/beiwe-backend#226 @biblicabeebli
The old wiki has been updated, every page has a notice stating the wiki is out of date, and a link pointing to the equivalent page on the new wiki.  The old wiki will no longer be updated.

Will not be closing issue until I have gone through every documentation page and updated or reviewed content for every page from the old wiki.
