### 2021-09-29 onnela-lab/beiwe-backend#226 @biblicabeebli
@zagorsky @hackdna @kwcarlso

I've updated and fixed most (all?) internal urls on the newly merged-in wiki content.  I would appreciate some help reviewing and updating it.

I can cover a lot, but there are areas that touch on the analysis side that I'm just not familiar with, and generally it needs some eyeballs.  Also, we should probably update it with respect to the existence of Forest.

I decided it was valuable to preserve the old structure, so there is now a [researcher's landing page](https://github.com/onnela-lab/beiwe-backend/wiki/%5BResearchers%5D-Home) and link to it from the [main landing page](https://github.com/onnela-lab/beiwe-backend/wiki).

Additionally, there is some interaction with the peer review work with respect to [this comment](https://github.com/onnela-lab/beiwe-backend/issues/210#issuecomment-929220831) on the peer review work.

cool final thing:
The wiki is its own git repo, which is super convenient for bulk changes like updating links after a file rename.  The contents of old wiki have been stuck into an `old-wiki` folder for ease of everything; for whatever reason the wiki continues to work flawlessly with this repo structure.
You can clone it with `git clone git@github.com:onnela-lab/beiwe-backend.wiki.git`
