### 2021-09-29 onnela-lab/beiwe-backend#208 @biblicabeebli
- Summary: - This is good and ticks all the boxes but I have one note. Are the specs on line 11 - 16 necessary or in the right place? It seems like it interrupts the sentences before and after.

I am not sure what you are referring to with respect to lines 11-16
