### 2021-09-29 onnela-lab/beiwe-backend#208 @biblicabeebli
I am requesting some review of documentation over on #226, and there has been a bunch of activity on #210.

This post seems like it is slightly higher level from the others, and I'm not totally sure what the distinction regarding "paper issues" is about.  Would appreciate a little bit more context and guidance.
