### 2021-09-29 onnela-lab/beiwe-backend#208 @erik-whiting
@biblicabeebli the "paper" we're referring to is the actual article that will be published in the journal and is located here: https://github.com/openjournals/joss-papers/blob/joss.03417/joss.03417/10.21105.joss.03417.pdf

Notes here are more like the traditional peer review process for scientific journals and they don't really relate to the software per se.
