### 2021-09-30 onnela-lab/beiwe-backend#226 @zagorsky
@biblicabeebli thanks for merging the wikis; I think that was a great move.  I thought it would be better to have one consolidated landing page for everyone (instead of one page for researchers, and one for engineers), so I merged all the links into one landing page on [the new wiki](https://github.com/onnela-lab/beiwe-backend/wiki).

I have two requests for you, since I don't have write access to the onnela-lab/wiki repo:
1. Could you please update the "This Wiki Has Moved" link to point at https://github.com/onnela-lab/beiwe-backend/wiki#documentation-for-researchers  ?
2. I think you should delete all content on each page of the old wiki, so nothing is left except the link pointing to the new page.  Otherwise, there's a risk that someone will update the old wiki, and then that info will eventually need to be reconciled with the new wiki, or it will eventually get lost when the old wiki gets deleted.

Thanks.  I don't think anyone else needs to review this; I did a cursory check that the links work and that the content is copied over.
