### 2021-09-30 d18e6af
Move data qty stats update to before unchunked files get deleted so if chunking fails and needs to rerun, so will data qty stats
