### 2021-09-30 249af20
Use SummaryStatisticDaily.update_or_create instead of bulk_update inside calculate data qty stats
