+++ libs/file_processing/data_qty_stats.py
@@ -38,20 +38,18 @@ def calculate_data_quantity_stats(
     for chunkregistry in chunkregistries:
         day = chunkregistry['time_bin'].astimezone(study_timezone).date()
         daily_data_qtys[day][chunkregistry['data_type']] += chunkregistry['file_size']
-    # Delete all existing DataQuantity objects for the participant
-    SummaryStatisticDaily.objects.filter(participant=participant).delete()
     # For each date, create a DataQuantity object
     data_qty_objects_to_be_created = []
     for day in daily_data_qtys:
         data_qty = {
             "participant": participant,
             "date": day,
+            "defaults": {}
         }
         for data_type in daily_data_qtys[day]:
             if data_type in ALL_DATA_STREAMS:
-                data_qty[data_type + '_bytes'] = daily_data_qtys[day][data_type]
-        data_qty_objects_to_be_created.append(SummaryStatisticDaily(**data_qty))
-    SummaryStatisticDaily.objects.bulk_create(data_qty_objects_to_be_created)
+                data_qty['defaults'][data_type + '_bytes'] = daily_data_qtys[day][data_type]
+        SummaryStatisticDaily.objects.get_or_create(**data_qty)
 
 
 def _utc_datetime_of_local_midnight_date(local_date, local_timezone):
