### 2021-09-30 0eccac1
Migration to populate historical data qty stats
