### 2021-09-30 8d24c48
Uncomment calls & texts summary stats in Tableau WDC API
