### 2021-10-05 2fba07b
[data-access-api-get-participants] built and tested in a browser, depending on exact details of behavior we may need to change the mimetype to just 'json' - it currently downloads the data when you hit the endpoint with a properly crafted url in a browser, json doesn't.
