### 2021-10-07 a859a1e
[fix-copy-study-relative-schedules] Python typo fix
