### 2021-10-07 7067d17
Update libs/file_processing/data_qty_stats.py
Replace defaultdict(lambda: 0) with defaultdict(int)

Co-authored-by: Eli Jones <biblicabeebli@gmail.com>
