### 2021-10-07 04ad491
Resolve code review issues
