+++ database/migrations/0061_historical_data_qty_stats.py
@@ -1,10 +1,36 @@
+from collections import defaultdict
 from django.db import migrations
 
-from database.user_models import Participant
-from libs.file_processing.data_qty_stats import calculate_data_quantity_stats
+from config.constants import ALL_DATA_STREAMS
 
 
 def populate_historical_data_qty_stats(apps, schema_editor):
+    ChunkRegistry = apps.get_model("ChunkRegistry")
+    Participant = apps.get_model("Participant")
+    SummaryStatisticDaily = apps.get_model("SummaryStatisticDaily")
+
+    def calculate_data_quantity_stats(participant: Participant):
+        """ Update the SummaryStatisticDaily  stats for a participant, using ChunkRegistry data """
+        study_timezone = participant.study.timezone
+        query = ChunkRegistry.objects.filter(participant=participant)
+
+        daily_data_quantities = defaultdict(lambda: defaultdict(int))
+        # Construct a dict formatted like this: dict[date][data_type] = total_bytes
+        for chunkregistry in query.values_list('time_bin', 'data_type', 'file_size'):
+            day = chunkregistry[0].astimezone(study_timezone).date()
+            daily_data_quantities[day][chunkregistry[1]] += chunkregistry[2]
+        # For each date, create a DataQuantity object
+        for day, day_data in daily_data_quantities.items():
+            data_quantity = {
+                "participant": participant,
+                "date": day,
+                "defaults": {}
+            }
+            for data_type, total_bytes in day_data.items():
+                if data_type in ALL_DATA_STREAMS:
+                    data_quantity["defaults"][f"{data_type}_bytes"] = total_bytes
+            SummaryStatisticDaily.objects.update_or_create(**data_quantity)
+
     for participant in Participant.objects.all():
         calculate_data_quantity_stats(participant)
 
