### 2021-10-07 ae40d3b
[sender-mismatch-exceptions] Updates to documentation, redocumented settings.py, purged some unused variables.
