+++ config/settings.py
@@ -1,59 +1,91 @@
 from os import cpu_count, getenv
 
 """
-On processing servers, instead of environment varables, append a line to config/remote_db_env.py,
-such as: os.environ['S3_BUCKET'] = 'bucket_name'
+Keep this document legible for non-developers, it is linked in ReadMe, and is the official
+documentation for all runtime parameters.
+
+On data processing servers, instead of environment varables, append a line to your
+config/remote_db_env.py file, formatted like this:
+    os.environ['S3_BUCKET'] = 'bucket_name'
+
+For options below that use this syntax:
+    getenv('REPORT_DECRYPTION_KEY_ERRORS', 'false').lower() == 'true'
+This means Beiwe is looking for the word 'true' and also accept "True", "TRUE", etc.
+If not provided with a value, or provided with any other value, they will be treated as false.
 """
 
-# for we retain the old access key in case someone deployed a super old EB server.
+# Credentials for running AWS operations, like retrieving data from S3 (AWS Simple Storage Service)
+#  This parameter was renamed in the past, we continue to check for the old variable name in order
+#  to support older deployments that have been upgraded over time.
 BEIWE_SERVER_AWS_ACCESS_KEY_ID = getenv("BEIWE_SERVER_AWS_ACCESS_KEY_ID") or getenv("S3_ACCESS_CREDENTIALS_USER")
 BEIWE_SERVER_AWS_SECRET_ACCESS_KEY = getenv("BEIWE_SERVER_AWS_SECRET_ACCESS_KEY") or getenv("S3_ACCESS_CREDENTIALS_KEY")
 
-# This is the secret key for the website. Mostly it is used to sign cookies. You should provide a
-#  cryptographically secure string to this value.
+# This is the secret key for the website, mostly it is used to sign cookies. You should provide a
+#  long string with high quality random characters. Recommend keeping it alphanumeric for safety.
 FLASK_SECRET_KEY = getenv("FLASK_SECRET_KEY")
 
-# The name of the s3 bucket that will be used to store user generated data, and backups of local
-# database information.
+# The name of the S3 bucket that will be used to store user generated data.
 S3_BUCKET = getenv("S3_BUCKET")
 
-# Domain name for the server
+# Domain name for the server, this is used for various details, and should be match the address of
+#  the frontend server.
 DOMAIN_NAME = getenv("DOMAIN_NAME")
 
 # A list of email addresses that will receive error emails. This value must be a comma separated
 # list; whitespace before and after addresses will be stripped.
 SYSADMIN_EMAILS = getenv("SYSADMIN_EMAILS")
 
-# Sentry DSNs (technically optional)
+# Sentry DSNs for error reporting
+# While technically optional, we strongly recommended creating a sentry account populating
+# these parameters.  Very little support is possible without it.
 SENTRY_DATA_PROCESSING_DSN = getenv("SENTRY_DATA_PROCESSING_DSN")
 SENTRY_ELASTIC_BEANSTALK_DSN = getenv("SENTRY_ELASTIC_BEANSTALK_DSN")
 SENTRY_JAVASCRIPT_DSN = getenv("SENTRY_JAVASCRIPT_DSN")
 
-# Developer debugging parameters for decryption issues
-REPORT_DECRYPTION_KEY_ERRORS = getenv('REPORT_DECRYPTION_KEY_ERRORS', 'false').lower() == 'true'
-STORE_DECRYPTION_KEY_ERRORS = getenv('STORE_DECRYPTION_KEY_ERRORS', 'false').lower() == 'true'
-STORE_DECRYPTION_LINE_ERRORS = getenv('STORE_DECRYPTION_LINE_ERRORS', 'false').lower() == 'true'
-
 # S3 region (not all regions have S3, so this value may need to be specified)
[…]
