### 2021-10-07 7f3b14e
[sender-mismatch-exceptions] more documentation fixes
