### 2021-10-07 bf66edc
[fix-copy-study-relative-schedules] avoid copying deleted surveys.
