### 2021-10-10 #3417 @erik-whiting
hi @fboehm, sorry for taking so long to respond. Eli responded to all my comments on the documentation issues, I will retest those items tomorrow (10/9) and update accordingly.

Regarding issues related to the paper (documented [here](https://github.com/onnela-lab/beiwe-backend/issues/208)), I don't think the person responsible for fixing those issues has seen my notes. Eli commented on that issue noting that he didn't really know what I was talking about. Maybe you could give the PI a nudge towards that issue so he can get the right person on it?
