### 2021-10-10 a234c1b
[fix-copy-study-relative-schedules] substantial refactor, code paths are now 100% shared except for exact behavior of how values are pulled from the request, and the notifications presented on the page on a success.
