+++ libs/copy_study.py
@@ -1,11 +1,13 @@
 import json
 from os import path
-from typing import Dict, List
+from typing import Dict, List, Union
 
-from flask import flash, request
-
+from constants.copy_study_constants import (ABSOLUTE_SCHEDULE_KEY, DEVICE_SETTINGS_KEY,
+    INTERVENTIONS_KEY, NEVER_EXPORT_THESE, RELATIVE_SCHEDULE_KEY, STUDY_KEY, SURVEY_CONTENT_KEY,
+    SURVEYS_KEY, WEEKLY_SCHEDULE_KEY)
 from database.common_models import JSONTextField
-from database.schedule_models import AbsoluteSchedule, Intervention, RelativeSchedule, WeeklySchedule
+from database.schedule_models import (AbsoluteSchedule, Intervention, RelativeSchedule,
+    WeeklySchedule)
 from database.study_models import Study
 from database.survey_models import Survey
 from libs.push_notification_helpers import repopulate_all_survey_scheduled_events
@@ -13,25 +15,46 @@ from libs.push_notification_helpers import repopulate_all_survey_scheduled_event
 
 NoneType = type(None)
 
-WEEKLY_SCHEDULE_KEY = "timings"  # predates the other schedules
-ABSOLUTE_SCHEDULE_KEY = "absolute_timings"
-RELATIVE_SCHEDULE_KEY = "relative_timings"
-INTERVENTIONS_KEY = "interventions"
 
-# keys used if various places, pulled out into constants for consistency.
-DEVICE_SETTINGS_KEY = 'device_settings'
-STUDY_KEY = 'study'
-SURVEY_CONTENT_KEY = 'content'
-SURVEY_SETTINGS_KEY = 'settings'
-SURVEYS_KEY = 'surveys'
+def unpack_json_study(json_string: str) -> Union[dict, List[str], List[dict]]:
+    """ Deserializes the data structure of a serialized study """
+    study_settings = json.loads(json_string)
+    device_settings = study_settings.pop(DEVICE_SETTINGS_KEY, {})
+    surveys = study_settings.pop(SURVEYS_KEY, [])
+    interventions = study_settings.pop(INTERVENTIONS_KEY, [])
+    return device_settings, surveys, interventions
 
-# "_id" is legacy, pk shouldn't occur naturally.  This list applies to Surveys and Studies.
-FIELDS_TO_PURGE = ('_id', 'id', 'pk',  'created_on', 'last_updated', 'object_id', "deleted")
+
+def format_study(study: Study) -> str:
+    """ Serializes a study, including surveys, their schedules, device settings, and interventions. """
+    device_settings = study.device_settings.as_unpacked_native_python()
+    purge_unnecessary_fields(device_settings)
+    return json.dumps(
+        {
+            SURVEYS_KEY: format_surveys(study),
+            DEVICE_SETTINGS_KEY: device_settings,
+            INTERVENTIONS_KEY: list(study.interventions.values_list("name", flat=True))
+        }
+    )
+
+
+def format_surveys(study: Study) -> List[dict]:
+    """ Serializes a survey and its schedule. """
+    surveys = []
+    for survey in study.surveys.filter(deleted=False):
+        # content, cleanup, then schedules.
+        survey_content = survey.as_unpacked_native_python()
+        purge_unnecessary_fields(survey_content)
+        survey_content[WEEKLY_SCHEDULE_KEY] = survey.weekly_timings()
+        survey_content[ABSOLUTE_SCHEDULE_KEY] = survey.absolute_timings()
+        survey_content[RELATIVE_SCHEDULE_KEY] = survey.relative_timings_by_name()
+        surveys.append(survey_content)
+    return surveys
 
 
[…]
