### 2021-10-10 #3417 @fboehm
Thanks @erik-whiting ! I just sent an email, and I'll paste the text of it here:

Hi, JP,

I'm writing with regards to the JOSS submission, Beiwe: A data collection platform for high-throughput digital phenotyping.

Your collaborator Eli has worked towards resolving the software issues that our reviewers have raised. There is an additional collection of issues that the reviewers raised about the paper that accompanies the submission. I just wanted to ensure that your entire team is aware of that collection of (paper-related) issues: https://github.com/onnela-lab/beiwe-backend/issues/208. We appreciate your attention to this.

Thanks again!

fred
