### 2021-10-19 #3417 @biblicabeebli
I'll ping my colleagues.  I've been cc'd on most of the relevant enails and I'm currently trying to bring up some internal infrastructure to pull topics stuck in email threads to more general awareness and visibility.

The general assessment that I'm not the one to address the paper issues directly is correct.
