### 2021-10-25 onnela-lab/beiwe-backend#255 @biblicabeebli
Drop Flask entirely for Django
Issue for tracking the Django porting process

## This is an involved process, the reasons are many, this work is long-overdue
This issue tracks the work on the branch [`eli/django`](https://github.com/onnela-lab/beiwe-backend/tree/eli/django)

This development work is critical for the platform health of Beiwe in general.  By ditching the old flask and adopting django especially in combination with the new tests that are part of this work, the substantial maintenance burden will (should) be reduced.

- Flask is a limited framework, it does not have the bulk features of Django.
- Flask is less widespread than Django, it will open up development opportunities to more people.
- The presence of Flask and Django causes general confusion, esp in bug reports and to other collaborators attempting to contribute.
- We shifted to using the Django ORM layer some time ago.  The remaining work consists of updating website endpoints to use the django prrimitives, and dropping any explicit flask dependencies.  Some work has involved developing shims to copy flask features.
- There is an occasional problem with database connection issues, this stems from a django-only feature that opens and closes the database connection intelligently based on incoming requests.  This behavior causes occasionally broken deployments that can only be diagnosed by checking for the number of open database connections on the RDS server, the only known fix is to reboot the database, or wait until the connections timeout.
- security - flask is missing many security features that need to be implemented externally to the web framework, this will reduce sysadmin work/responsibility as we can finally automat a couple features.
- tests - building tests has long been problematic for the platform.  Running django-native allows a much easier test framework that the developers are already familiar with.  it also means we don't have to maintain our own weird compatibility layer in the tests.
- general dependency reduction.

## Current Status
- all _website_ endpoints have at least a first-pass transformation to using django, substantial testing is needed but many things load.  Quality is any of Alpha, Beta, or Release Candidate, depending on ease of testing, mostly Beta+.
- API endpoints are mostly Alpha, endpoints that are required for loading pages are RC.
- admin/researcher authentication is pretty good, Beta or RC..
- download api authentication needs testing and development, Alpha.
- tableau and its authentication is **not implemented**.
- mobile pages and device authentication are Alpha, first pass with zero testing.

## Tests
Tests have been moved to a top-level folder.  We are currently using the djangotestrunner, open to using additional libraries.
- we don't need 100% test code coverage of all endpoints, but all endpoints must have at least one test on them.
- all authentication code must have **100% code coverage before release.**
- all device endpoints must have **100% code coverage before release.**
- selenium tests will not be included in this work.
- javascript - no plans at this time, **looking for input**.
- the `factory_boy` dependency is currently in use only on the (preexisting) tableau API tests.  **Use of this library is currently considered an impediment to testing because it actively obscures the database aspects of the tests, making correctness difficult to determine.**  Further discussion on this topic is expected.

## How To Test Code Coverage Using The `coverage` Library
- you will need to pip install the `coverage` library into your project python environment.  There is a new requirements_testing.txt file.
- `coverage` is fairly simple to use, but it does generate files and you may forget to clean up after previous runs, here is a bash function that will allow you to easily run tests with coverage without having to keep details about it in your head.
- use the below function like this, documentation is inlined.  First parameter is the folder to run `coverage` over, all following parameters get passed into the test command.

`your_renamed_function authentication tests.test_endpoints.TestAuthentication`

```
function please_rename_me(){
    if (( $# != 2 ))
        then echo "need 2 params: folder to apply python coverage, and then tests to run (empty string for all tests)"
        return
    fi

    coverage erase  # erase existing coverage content
    rm -rf htmlcov  # delete the generated html report

    # coverage requires a folder target containing all python code you want the coverage report on to operate on.
    #   provide "." to run on all code, but that can be slow.
    # Executing a test with coverage takes longer than normal, recommend using the time directive to give you
    #   sanity-saving time feedback so you can adapt your workflow.
    # The testing directive here is to by-default run with --keepdb for speed, and to run in extra verbose mode because otherwise print statements don't work and that can be aggrivating.  Change it if you dislike.

    # Capture the first parameter you passed to this function
    # `shift` `shifts all of this function's parameters down by one, deletes the contents of $1 that we grabbed above.
    #   The remaining parameters get passed directly to `manage.py test`, which is populated with some defaults that you should review.
    COVERAGE_FOLDER=$1
    shift

    time coverage run --source="$COVERAGE_FOLDER" manage.py test -v2 --keepdb $*

    # (this can take some time, so I've stuck a print statement)
    echo "\n" building html report
    coverage html

    # attempts to open the new coverage file in a way that works on mac and linux, your call whether to include.
    echo Opening...
    open -a '/Applications/Google Chrome.app/' htmlcov/index.html 2> /dev/null || xdg-open htmlcov/index.html
}
```
