### 2021-10-30 #3417 @fboehm
> JP has just posted his responses over on the paper issues post in the beiwe-backend repository. He sent me a couple comments and questions:
>
> He says he replied the way he would typically reply to blind or double blind review comments. Let us know if it needs to be of a different form in the comment threads. (I can't think of a reason why this wouldn't be fine?)
>

Yes, that should be fine.

> We are not sure where to provide the modified manuscript. I've run through the early sections of this thread, I've seen that there is an at-whedon bot-thing for doing some editorial actions, and I've seen the [Joss Papers tree](https://github.com/openjournals/joss-papers/tree/joss.03417/joss.03417) location of the current pdf. Coincident with this is the question of the change tracking mechanism for JOSS - are the changes tracked by a git repository containing the paper content, or is there a different process. (I'll try and get through some documentation on this and work it out for JP, but if you have the links handy that would save me some time.)

You'll want to add the paper.md, ie, the markdown version of the paper, to your repository. Most authors add it to the main branch, but some prefer to add it to an additional branch. You shouldn't add a pdf version of the markdown file to the repository. Please let me know if that addresses your questions. Here's a link to some of the relevant documentation: https://joss.readthedocs.io/en/latest/submitting.html
