### 2021-10-30 #3417 @fboehm
@biblicabeebli - it looks like your repository's current placement of the paper.md is correct: https://github.com/onnela-lab/beiwe-backend/blob/ef28d952b7158c7f03df263003f05824e7909d1b/paper.md.
So you'll just want to replace it with the updated version of paper.md.
