### 2021-10-31 #3417 @Tam-Pham
Hi all,

Firstly, I would like to apologize for the delay on my part. As I'm not familiar with server applications, it took me some time to complete my review on the functionality part. With the revisions made to the documentation of the development setup, I have managed to deploy Beiwe-backend locally and hence have ticked off all the related checkboxes.

I have not ticked the following checkboxes:

-  **Contribution and authorship**: it's mentioned [here](https://github.com/onnela-lab/beiwe-backend/issues/208#issuecomment-954128235) about a revision in authorship and thus I will wait for the updated changes in the `paper.md` before completing this point.
- **Automated tests**: please correct me if I'm wrong but the software doesn't seem to have an automated test system (e.g., GH actions, Travis, CircleCI...) set up. Could the authors clarify on how the functionality of the software can be verified?
- **State of the field**: I will wait for the updated changes that address the comments [here](https://github.com/onnela-lab/beiwe-backend/issues/208#issuecomment-954128235) before completing this point.
