### 2021-11-03 onnela-lab/beiwe-backend#255 @biblicabeebli
(Notification because I am mentioning this branch to other people now that it has good progress)

There has been a regression in functionality, I screwed up slightly on the exact implementation of the `url` jinja template extension, so a lot of `href="{{ url ... }} "` things are currently broken.  I'm hitting them and fixing them as I go through building tests.
