### 2021-11-03 #3417 @biblicabeebli
(ack, meant to comment yesterday)
> Automated tests: please correct me if I'm wrong but the software doesn't seem to have an automated test system (e.g., GH actions, Travis, CircleCI...) set up. Could the authors clarify on how the functionality of the software can be verified?

With respect to tests, the fact is we have very few.  However, I am in the process of swapping out the base web framework, Flask, for Django, and writing a [hopefully comprehensive!] test suite.

This backend framework port, and the coincident work on tests, is tracked here: https://github.com/onnela-lab/beiwe-backend/issues/255

If you check out onto the `eli/django` branch you will be able to log in and many of the pages currently load - except I made some mistakes in applying a particular bulk change to the jinja templates (HTML templating language) so there are more rendering errors than there were a week or two ago.

To run the webserver locally you no longer run `python app.py`, instead you run `python manage.py runserver`.  Note that requirements.txt is not final, in fact I don't even recall if it has been updated.  There is a new additional requirements file for testing.  There are directions in the post linked above for running tests along with the `coverage` library to get insight on how much of the code is touched by tests.

I intend to configure a Github CI at the end of this work, and a pull request requirement that all tests pass.
