### 2021-11-03 #3417 @erik-whiting
@Tam-Pham I interpreted the "or manual steps described so that the functionality of the software can be verified" to be satisfied by the existing documentation. I saw that there were tests and even though they weren't automated, I thought the documentation (after it was fixed) qualified as steps for verifying functionality. That's just me though obviously, it's your review. Just noting why I noted that checklist item as good on my review.
