### 2021-11-09 onnela-lab/beiwe-backend#255 @biblicabeebli
Udpate for those paying attention:
The Django port is at roughly "Pretty much done." on the [Done Scale](http://20px.com/blog/2014/12/05/done-scale/).

- Every use of the Flask library has been removed from active code, Django can now run the server and every endpoint is present, many of them even work.
- References to Flask are still present, and no work has been done to match configuration present in `app.py`.
- There are now 50 tests of varying quality covering quite a few of the api and pages files.  I'm simply going down through the `urls` file and covering every endpoint.
- I have used `coverage` on a handful of the tests, I will try using it to fill in tests in more detail later on for critical items.  The goal right now is to cover every endpoint with the test client with a reasonable amount of confirmation that behavior is as intended, and that pages load and return the correct status codes.
- There are no unit tests, and no tests for code that is not directly served from the Flask app.  Such tests are not expected to be present before merging into `main`.
- I have scattered quite a few `FIXME`'s across the codebase to address many issues that should be cleaned up before merging to `main`, but only some of those are blockers.
- No testing work has been completed on any of the `mobile_api` code.

The testing abstractions are still in development, but the major patterns at play, excluding `mobile_api`, have been identified.  Documentation of the abstractions is pretty sparse as they could still change, but I will get around to documenting them.
