### 2021-11-10 #3417 @jponnela
Hi everyone (@erik-whiting, @fboehm, @Tam-Pham, @biblicabeebli),

I've just updated paper.md and paper.bib files to reflect the responses I submitted about two weeks ago. I believe and hope that we've addressed all the points that were raised, but please let us know if we missed anything or if there are additional steps we should take.
