### 2021-11-12 onnela-lab/beiwe-backend#208 @erik-whiting
Thanks for addressing everything @jponnela, the updates look great. Closing this and updating my review at https://github.com/openjournals/joss-reviews/issues/3417
