### 2021-11-12 #3417 @erik-whiting
@fboehm I've reread the Beiwe paper and believe @jponnela and the rest of the team has addressed everything I pointed out. The paper is great and @biblicabeebli documentation updates (completed a while ago) have already addressed all the non-paper issues I brought up. I've checked off the remaining items in my checklist and have closed the issues I opened on their repository.

My review is complete and I say this is ready for publication.
