### 2021-11-13 #3417 @fboehm
@jponnela  and @biblicabeebli - here are a few minor corrections to paper.md that are needed before we proceed. Line numbers refer to those in the pdf.

lines 57-66: use lower case letters to start the words after the semicolons.
line 65: change (7) to (8).
line 78: fix citation for EARS. I *think* that you can fix this by deleting line 260 of paper.bib.
line 149: delete "a", i.e., "encrypt a symmetric" should be "encrypt symmetric".

Please let me know, by "@" mentioning me, fboehm, once these minor changes are incorporated.

Thanks again!
