### 2021-11-15 onnela-lab/beiwe-backend#255 @biblicabeebli
### Udpate!
The Django port is at roughly "Done except for..." on the Done Scale.

- There are now 107 tests.
- _Many_ Django port issues have been resolved.
- Many bugs have been fixed.
- There are currently 57 `fixme`s across the codebase, about half of them attached to the existing tests noting additional cases to be implemented.
- No real update at this time with respect to use of Coverage to gauge the quality of tests.
- Have not gotten to tests for data download, tableau, or mobile api.  Those are the "hard ones", but fortunately all the expected changes have been made and "only the bugs" remain.
- The abstractions and code tools needed to make concise, legible tests for HTTP endpoints are largely complete.
- The abstractions are mostly documented!
- The tools for generating appropriate database objects for testing purposes are still growing, they are located in `tests/helpers.py`.
- Requirements have been updated, testing requirements have been updated.
- The branch has 139 commits and is up-to-date with code changes on `main`.

Goal is to get this on our staging server before Thanksgiving day in the US (November 25).
