### 2021-11-17 #3417 @jponnela
Thank you for pointing these out @fboehm, should be all fixed now.

@biblicabeebli, I created JP/paper branch for implementing these changes.

Thank you!
