### 2021-11-18 #3417 @fboehm
Thanks, @jponnela and @biblicabeebli !

The next step is for you to make an archive of the repository with a service like zenodo and to make a new github release. Please share here the doi from zenodo and the version number of the release. Please let me know if you have any questions! Thanks again!
