### 2021-11-19 onnela-lab/beiwe-backend#255 @biblicabeebli
Moderate progress this week, but the tedious part of bulk test building and "does page render without crashing" are finished. 154 commits total.

I've built tests for every endpoint that is part of the normal researcher login session, moving on to the endpoints that require alternate credentials, or come from the mobile apps.

Done scale progress: unchanged, still "Done except for..."
