### 2021-11-19 #3417 @biblicabeebli
@fboehm I'll take care of that.  Is there a timeliness factor at play here?  Next week is complicated by the thanksgiving holiday and I'm trying to manage expectations on several things.   :b
