### 2021-11-29 onnela-lab/beiwe-backend#255 @biblicabeebli
Mobile api now has test coverage.
- The ios endpoints are weird, they use http basic auth for bad reasons, and I have not worked out how to test that detail yet, but other than weird authentication behavior they are literally the same code.
- Device upload does not yet have a success testcase, that is particularly hairy code to "just build" valid data for, so I will finish that test when I get to real device testing, which will be... soon.  (This week? :D )

@alvinsiu I have ported your tests for tableau stuff, thanks for putting those together.  They may be unrecognizable.

The Massive 2,415 line test_endpoints.py file will be factored out into some yet-to-be-determined organization, I have avoided doing this because imho there are a **lot** of endpoints that should also be moved around, and I have implemented two meta tests to help me track these organizational details.

### You can now just run all tests and the experience is reasonable.
- TESTS PASS AND THAT ACTUALLY MEANS SOMETHING!
- Test coverage in `api` and `pages` are 69% and 76%.  This will increase some as the final tests are put together and some fixmes are addressed in the existing tests.  100% coverage is not a goal at this time.  `libs` only has 28% coverage, but it contains developer utils not accessed in the live codebase, and components that are unreachable from the frontend.
- The two aforementioned meta tests are currently skipped (they do not pass right now).
- All tests can be run in parallel, just pass in `--parallel number_of_cpu_cores`.  The test architecture is expected to maintain this to allow for a fast iteration process for developers in the future, and is recommended.
- Strongly recommend running with `--keepdb` parameter, there are 62 migrations.
- There is extra terminal output on verbosity level 2 or greater.  Output from tests are bracketed with `==` and new lines. Calls to the messages library (e.g. `messages.error`, etc.) will additionally print their output to the terminal to assist with test development.
- Calls that would hit s3 in the `libs.s3` functions will automatically fail with a(n insane) Boto3 error.  This is accomplished by patching the Bucket name to be a None.  This is the case for all tests and cannot be disabled.  Tests that require S3 functionality must use the mock library's patch functionality.  It is not a goal right now to substantially test `libs.s3`.
- requirements have been updated, and a requirements_testing.txt file is present that includes the `mock` and `coverage` libraries, as well as `tblib`, which is necessary to serialize traceback objects when they occur in parallel mode.
- All work has been to be compatible with Pythons 3.6, 3.7, and 3.8.  **Note that the current `main` branch is not compatible with 3.8**

@zagorsky given the sheer breadth of the work and the presence of the new tests I will not be asking for intensive review, but this work is substantially progressed and I intend to have it on at least the staging server this week.  Please take a look at the branch, you may want a new python env, and a new python version.  As soon as this is done (or at least on staging for testing) I will be starting on the Python 3.8 platform upgrade work.  I have not looked at the work on the messaging branch at all.  If you want to take a whack at that merge/rebase you should do it as a merge into [a copy of] the `eli/django` branch; the rebase is painful, there are moved and renamed files, import conflicts, copy-pastes of code, etc.
