### 2021-11-30 8666722
Add full local setup with docker
- Add separate configuration for s3 endpoint
- Add dockerfile to build backend image
- Add docker-compose with minio and rabbitmq
- Add MacOS setup guide
