### 2021-12-09 onnela-lab/beiwe-backend#255 @biblicabeebli
Update:
Done scale: Done v3
We are live on our staging server.
iOS support is broken.
Everything else seems to work.
