### 2021-12-09 onnela-lab/beiwe-backend#255 @biblicabeebli
and now iOS should work.  Unless I've misunderstood something about http basic auth
