### 2021-12-09 #3417 @jponnela
Hi @fboehm and @biblicabeebli! The DOI is 10.5281/zenodo.5758811. I called the version number of the release v1.0.0.
