### 2021-12-10 #3417 @fboehm
@jponnela @biblicabeebli - once the archive title is updated, the submission will then be handed off to an associate editor in chief. They'll do the final checks before publication.
