### 2021-12-13 4758f83
Bump django from 2.2.24 to 2.2.25
Bumps [django](https://github.com/django/django) from 2.2.24 to 2.2.25.
- [Release notes](https://github.com/django/django/releases)
- [Commits](https://github.com/django/django/compare/2.2.24...2.2.25)

---
updated-dependencies:
- dependency-name: django
  dependency-type: direct:production
...

Signed-off-by: dependabot[bot] <support@github.com>
