### 2021-12-17 #3417 @jponnela
Thank you so much everyone, my first time publishing in JOSS and I'm very impressed!

@Tam-Pham @erik-whiting @fboehm @kyleniemeyer @biblicabeebli
